// plugins/main/server/lib/generate-alerts/types.ts
export interface Rule {
  id: string;
  level: number;
  description: string;
  groups: string[];
  firedtimes?: number;
  mail?: boolean;
  gdpr?: string[];
  pci_dss?: string[];
  tsc?: string[];
  hipaa?: string[];
  nist_800_53?: string[];
  nist_800_83?: string[];
  gpg13?: string[];
  pci?: string[];
  frequency?: string;
  info?: string;
}